\name{Internal functions}
\alias{MLDirichletPriors}
\alias{classifyProbes}
\alias{rdirichlet}
\alias{proportions}
\alias{find.MCR}
\alias{beta}
\alias{findChangePoints}
\alias{equivalenceClasses}
\alias{propsList}
\alias{rowCount}
\alias{MLalphas}
\alias{logBetaOverBeta}
\alias{cytobandIds}
\alias{summariseRegions}
\alias{logsubgamma}
\alias{cytobands}
\alias{ClassifySegList-class}
\alias{dim,ClassifySegList-method}
\alias{dimnames,ClassifySegList-method}
\alias{show,ClassifySegList-method}
\alias{length,ClassifySegList-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal functions for the bayesCGH package}
\description{
These are internal functions, generics, classes and data for the
bayesCGH package that will rarely if ever be explicitly called by the
user. If you need to call them, and are unable to resolve any issues,
please contact the maintainer of this package.
}

\author{Thomas Hardcastle}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}

