\name{writeSummaries}
\alias{writeSummaries}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Writes summaries produced by 'summariseSelection' function to file.}
\description{
Writes summaries produced by
  \code{\link[bayesCGH:ClassifySegList]{summariseSelection}} to a set of
  tab-delimited files.
}
\usage{
writeSummaries(data, dir = ".", table.name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{Output from
    \code{\link[bayesCGH:summariseSelection]{summariseSelection}} function.}
  \item{dir}{Directory to write to.}
  \item{table.name}{Table name from which filenames for tab-delimited
    files are acquired.}
}
\value{
  Set of tab-delimited files containing summary data generated by
  bayesCGH package.
}
\author{Thomas Hardcastle}
\seealso{\code{\link[bayesCGH:summariseSelection]{summariseSelection}}}

\keyword{IO}
